Do not add the build.sh scripts to global scripts

+++ btcpay-setup.sh
@@ -259,9 +259,12 @@ fi
 cd $ORIGINAL_DIRECTORY
 
 for scriptname in *.sh; do
+    if [ "$scriptname" == "build.sh" -o "$scriptname" == "build-pregen.sh" ] ; then
+        continue;
+    fi
     echo "Adding symlink of $scriptname to /usr/bin"
     chmod +x $scriptname
     rm /usr/bin/$scriptname &> /dev/null
-    ln -s $(pwd)/$scriptname /usr/bin
+    ln -s "$(pwd)/$scriptname" /usr/bin
 done
