+++ btcpay-update.sh
@@ -7,9 +7,12 @@ set -e
 cd "$BTCPAY_BASE_DIRECTORY/btcpayserver-docker"  
 git pull --force
 # Generate the docker compose in BTCPAY_DOCKER_COMPOSE
-. ./build.sh
+#. ./build.sh
 
 for scriptname in *.sh; do
+    if [[ "$scriptname" == "build.sh" ] || [ "$scriptname" == "build-pregen.sh" ]] ; then
+        continue;
+    fi
     echo "Adding symlink of $scriptname to /usr/bin"
     chmod +x $scriptname
     rm /usr/bin/$scriptname &> /dev/null
